refactor RULE_SEVERITY_OPTIONS
